[refactor-main] add mop files and update main
